+++ core/data/src/main/java/dev/anilbeesetti/nextplayer/core/data/mappers/MediumEntityToVideo.kt
@@ -1,19 +1,8 @@
 package dev.anilbeesetti.nextplayer.core.data.mappers
 
 import dev.anilbeesetti.nextplayer.core.common.Utils
-import dev.anilbeesetti.nextplayer.core.data.models.VideoState
 import dev.anilbeesetti.nextplayer.core.database.entities.MediumEntity
 import dev.anilbeesetti.nextplayer.core.model.Video
-
-fun MediumEntity.toVideoState(): VideoState {
-    return VideoState(
-        path = path,
-        position = playbackPosition,
-        audioTrack = audioTrackIndex,
-        subtitleTrack = subtitleTrackIndex,
-        playbackSpeed = playbackSpeed
-    )
-}
 
 fun MediumEntity.toVideo() = Video(
     id = mediaStoreId,
@@ -28,6 +17,5 @@ fun MediumEntity.toVideo() = Video(
     size = size,
     dateModified = modified,
     formattedDuration = Utils.formatDurationMillis(duration),
-    formattedFileSize = Utils.formatFileSize(size),
-    thumbnail = null
+    formattedFileSize = Utils.formatFileSize(size)
 )
